+++ Coursework/domain/src/main/resources/db/flyway/V20240322222706__create_shows_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE shows
 (
     id SERIAL PRIMARY KEY,
-    price INT NOT NULL,
+    price DECIMAL NOT NULL,
     start_time TIMESTAMP NOT NULL,
     end_time TIMESTAMP NOT NULL,
     movie_id INT NOT NULL,
